updated tables from the database to reply the correct data structure

+++ backend/database/migrations/2025_08_11_171732_create_forecasts_table.php
@@ -21,7 +21,7 @@ return new class extends Migration {
             $t->unsignedTinyInteger('month');          // 1..12
 
             // Valor pronosticado (entero, sin decimales)
-            $t->unsignedBigInteger('volume');
+            $t->Integer('volume');
 
             // Versionado (controlás lógica en controller/frontend)
             $t->unsignedInteger('version')->default(1);
